replaced browserify with webpack and configured yoastseo as a external
