--- grunt/config/browserify.js
@@ -1,8 +1,0 @@
-module.exports = {
-	build: {
-		files: {
-			"js/yoastseo-woo-plugin-590.js": [ "js/src/yoastseo-woo-plugin.js" ],
-			"js/yoastseo-woo-replacevars-590.js": [ "js/src/yoastseo-woo-replacevars.js" ],
-		},
-	},
-};
