+++ package.json
@@ -28,12 +28,10 @@
     "eslint-plugin-react": "^7.7.0",
     "eslint-plugin-yoast": "^1.0.1",
     "grunt": "^1.0.2",
-    "grunt-browserify": "^5.0.0",
     "grunt-checktextdomain": "^0.1.1",
     "grunt-contrib-clean": "~0.6.0",
     "grunt-contrib-compress": "^1.4.3",
     "grunt-contrib-copy": "^1.0.0",
-    "grunt-contrib-uglify": "^0.10.0",
     "grunt-contrib-watch": "^0.6.1",
     "grunt-eslint": "^20.1.0",
     "grunt-glotpress": "^0.2.1",
@@ -42,9 +40,14 @@
     "grunt-phplint": "0.0.5",
     "grunt-replace": "^0.8.0",
     "grunt-shell": "~1.1.1",
+    "grunt-webpack": "^3.1.2",
     "grunt-wp-i18n": "^0.4.8",
     "load-grunt-config": "^0.19.2",
-    "time-grunt": "^1.0.0"
+    "lodash.defaultsdeep": "^4.6.0",
+    "time-grunt": "^1.0.0",
+    "unminified-webpack-plugin": "^2.0.0",
+    "webpack": "^4.16.2",
+    "webpack-cli": "^3.1.0"
   },
   "yoast": {
     "pluginVersion": "7.9"
