+++ tools/grunt/config/checktextdomain.js
