+++ tools/grunt/config/clean.js
@@ -1,0 +1,4 @@
+module.exports = {
+	jsDistFiles:
+		['js/*.js', '!js/src/*.js']
+};
