+++ tools/grunt/config/jsonlint.js
