+++ tools/grunt/config/set-version.js
