+++ tools/grunt/config/webpack.js
@@ -1,0 +1,8 @@
+/* global require */
+const webpackConfigProd = require( "../../webpack/webpack.config.prod" );
+const webpackConfigDev = require( "../../webpack/webpack.config.dev" );
+
+module.exports = {
+	buildDev: webpackConfigDev,
+	buildProd: webpackConfigProd,
+};
