+++ tools/webpack/webpack.config.dev.js
@@ -1,0 +1,17 @@
+/* global require, module */
+const defaults = require( "./webpack.config.default" ).defaults;
+const webpack = require( "webpack" );
+
+const devConfig = {
+	mode: "development",
+	devtool: "eval",
+	plugins: [
+		new webpack.DefinePlugin( {
+			"process.env": {
+				NODE_ENV: JSON.stringify( "development" ),
+			},
+		} ),
+	]
+};
+
+module.exports = defaults( devConfig );
